Update style.css
I really like the simplistic layout and rich information of your homepage. The font and background images also looks really good. One thing that I can think of is that the margin for each portfolio item can be a little bigger to make it more comfortable for users to look at. This is a great homepage!